More correct type handling of resource subscript rewriting
* We now check the intrinsic function parameter types against the input types to ensure we pick the correct intrinsic function id
* We now only search for functions in the relevant object type to ensure we pick the function id for the correct object
* We now correctly cast uint -> int for scalar locations
* StructuredBuffer Load now use an int location which is what the intrinsic it is generating for expects

+++ msl/tests/msl_export_tests.rs
@@ -1318,7 +1318,7 @@ struct StructuredBuffer
     device const T* address;
     uint64_t size;
 
-    T Load(uint location) const {
+    T Load(int location) const {
         return address[location];
     }
 };
@@ -1336,7 +1336,7 @@ void sub_static(thread uint& s_globalState) {
 }
 
 void sub_buffer(const helper::StructuredBuffer<uint4> g_buffer) {
-    g_buffer.Load(0u);
+    g_buffer.Load((int)0u);
 }
 
 void sub_groupshared(threadgroup float& lds_x) {
